requirements.txt
keyboard
pyinstaller
customtkinter
pillow
mouse